+++ packages/vtable/src/core/BaseTable.ts
@@ -4260,8 +4260,24 @@ export abstract class BaseTable extends EventTarget implements BaseTableAPI {
     this.render();
     const stage = this.scenegraph.stage;
     if (stage) {
-      const contentWidth = this.tableX + this.getAllColsWidth();
-      const contentHeight = this.tableY + this.getAllRowsHeight();
+      let contentWidth = this.tableX + this.getAllColsWidth();
+      let contentHeight = this.tableY + this.getAllRowsHeight();
+      if (this.internalProps.legends) {
+        this.internalProps.legends.forEach(legend => {
+          if (legend.orient === 'right') {
+            contentWidth = Math.max(contentWidth, legend.legendComponent.globalAABBBounds.x2);
+          } else if (legend.orient === 'bottom') {
+            contentHeight = Math.max(contentHeight, legend.legendComponent.globalAABBBounds.y2);
+          }
+        });
+      }
+      if (this.internalProps.title) {
+        if (this.internalProps.title._titleOption.orient === 'right') {
+          contentWidth = Math.max(contentWidth, this.internalProps.title.getComponentGraphic().globalAABBBounds.x2);
+        } else if (this.internalProps.title._titleOption.orient === 'bottom') {
+          contentHeight = Math.max(contentHeight, this.internalProps.title.getComponentGraphic().globalAABBBounds.y2);
+        }
+      }
       if (contentWidth >= this.canvasWidth && contentHeight >= this.canvasHeight) {
         stage.render();
         const buffer = stage.window.getImageBuffer(type);